Move stm (and clone) specific io initialisation to platform
