+++ src/platform/APM32/mk/APM32F4.mk
@@ -147,6 +147,7 @@ endif
 
 MCU_COMMON_SRC = \
         common/stm32/system.c \
+        common/stm32/io_impl.c \
         APM32/startup/system_apm32f4xx.c \
         drivers/inverter.c \
         drivers/dshot_bitbang_decode.c \
