+++ src/platform/STM32/mk/STM32_COMMON.mk
@@ -4,9 +4,12 @@ INCLUDE_DIRS += $(PLATFORM_DIR)/common/stm32
 MCU_COMMON_SRC += \
             common/stm32/bus_spi_pinconfig.c \
             common/stm32/bus_spi_hw.c \
+            common/stm32/io_impl.c
 
 SIZE_OPTIMISED_SRC += \
             common/stm32/bus_spi_pinconfig.c
 
 SPEED_OPTIMISED_SRC += \
-            common/stm32/bus_spi_hw.c
+            common/stm32/bus_spi_hw.c \
+            common/stm32/io_impl.c
+
